requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
idna==3.2
idna-ssl==1.1.0
multidict==5.2.0
typing-extensions==3.10.0.2
wincertstore==0.2
yarl==1.6.3